Refactor: footer데이터 map함수 활용과 리뷰반영 - footer데이터를 상수데이터+map함수로 반복되는 코드제거 - 로그인페이지에 변수명이 길어 해당 구조분해할당이 비효율적이라 삭제
