+++ src/pages/seyeonPark/Main/MainComponent/footerData.js
@@ -1,0 +1,14 @@
+const INFO = [
+  { id: 1, content: '소개' },
+  { id: 2, content: '도움말' },
+  { id: 3, content: 'API' },
+  { id: 4, content: '채용정보' },
+  { id: 5, content: '개인정보처리방침' },
+  { id: 6, content: '약관' },
+  { id: 7, content: '위치' },
+  { id: 8, content: '인기계정' },
+  { id: 9, content: '해쉬태그' },
+  { id: 10, content: '언어' },
+];
+
+export default INFO;
